requirements.txt
rdflib
pytest
geopandas
validators
requests
python-dotenv